Move ShapeValidator into a sub-namespace
